# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitPulse Fitness Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: getWorkouts
      tags: [workouts]
      summary: Get user workouts
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "201":
          description: Workout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get a single workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    put:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkoutRequest"
      responses:
        "200":
          description: Workout updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Workout:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        workoutName:
          type: string
        category:
          type: string
        duration:
          type: integer
        caloriesBurned:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
        notes:
          type: string
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, userId, workoutName, duration, caloriesBurned]
    WorkoutExercise:
      type: object
      properties:
        exerciseId:
          type: integer
        exerciseName:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        duration:
          type: integer
      required: [exerciseName]
    CreateWorkoutRequest:
      type: object
      properties:
        userId:
          type: integer
        workoutName:
          type: string
        category:
          type: string
        duration:
          type: integer
        caloriesBurned:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
        notes:
          type: string
        completedAt:
          type: string
          format: date-time
      required: [userId, workoutName, duration, caloriesBurned]
    UpdateWorkoutRequest:
      type: object
      properties:
        workoutName:
          type: string
        duration:
          type: integer
        caloriesBurned:
          type: integer
        notes:
          type: string
...