# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitPulse Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutExercise } from "./workoutExercise";

export interface Workout {
  id: number;
  userId: number;
  workoutName: string;
  category?: string;
  duration: number;
  caloriesBurned: number;
  exercises?: WorkoutExercise[];
  notes?: string;
  completedAt?: Date;
  createdAt?: Date;
}